fastboot::erase: skip pointless erasing of verity partitions.
when a partition runs verity (system|vendor), erasing it is pointless since the
formating with the next image is already subject to dm-verity validation and possibly
fec correction (if enabled).  so the purpose of erasing those partitions to avoid
error or tampering with the following formating is moot.

skipping those situations allows to speed up the overall fastboot flashing command
(when [-u] is not in use) and|or flashall command.

refs #SWANDROID-4735
Bug:70854486

Change-Id: I2f221769cc2cf825579db7561154200ce8201cc2
Signed-off-by: Pierre Couillaud <pierre.couillaud@broadcom.com>